remove conflicting types, refactor `utils` to have `constants` and `calculations` modules
