+++ src/data/index.ts
@@ -1,3 +1,2 @@
-export * from './constants';
 export * from './poolList';
 export * from './tokenList';
